+++ lib/validator_regex_pattern.dart
@@ -2,5 +2,6 @@ class ValidatorRegexPattern {
 
   static const String email = r'^[^@]+@[^@]+\.[^@]+$';
   static const String url =  r'^(https?|ftp)://[^\s/$.?#].[^\s]*$';
+  static const String signedDecimalValidator =  r'^-?\d+(\.\d+)?$';
 
 }